test: add tests to useCase of updateBoardPhase
