+++ backend/src/modules/boards/applications/update-board-phase.use-case.spec.ts
@@ -1,0 +1,139 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { ConfigService } from '@nestjs/config';
+import { EventEmitter2 } from '@nestjs/event-emitter';
+import * as CommunicationsType from 'src/modules/communication/interfaces/types';
+import * as Boards from 'src/modules/boards/interfaces/types';
+import { boardRepository } from '../boards.providers';
+import { BoardRepositoryInterface } from '../repositories/board.repository.interface';
+import { SendMessageServiceInterface } from 'src/modules/communication/interfaces/send-message.service.interface';
+import { BoardFactory } from 'src/libs/test-utils/mocks/factories/board-factory.mock';
+import { BoardPhases } from 'src/libs/enum/board.phases';
+import { TeamFactory } from 'src/libs/test-utils/mocks/factories/team-factory.mock';
+import { SLACK_ENABLE, SLACK_MASTER_CHANNEL_ID } from 'src/libs/constants/slack';
+import { FRONTEND_URL } from 'src/libs/constants/frontend';
+import { UpdateFailedException } from 'src/libs/exceptions/updateFailedBadRequestException';
+import { DeepMocked, createMock } from '@golevelup/ts-jest';
+import { UseCase } from 'src/libs/interfaces/use-case.interface';
+import { UpdateBoardPhaseUseCase } from './update-board-phase.use-case';
+import { BoardPhaseDto } from 'src/libs/dto/board-phase.dto';
+
+describe('UpdateBoardPhaseUseCase', () => {
+	let useCase: UseCase<BoardPhaseDto, void>;
+
+	let boardRepositoryMock: DeepMocked<BoardRepositoryInterface>;
+	let eventEmitterMock: DeepMocked<EventEmitter2>;
+	let configServiceMock: DeepMocked<ConfigService>;
+	let slackSendMessageServiceMock: DeepMocked<SendMessageServiceInterface>;
+
+	beforeAll(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			providers: [
+				UpdateBoardPhaseUseCase,
+				{
+					provide: CommunicationsType.TYPES.services.SlackSendMessageService,
+					useValue: createMock<SendMessageServiceInterface>()
+				},
+				{
+					provide: boardRepository.provide,
+					useValue: createMock<BoardRepositoryInterface>()
+				},
+				{
+					provide: EventEmitter2,
+					useValue: createMock<EventEmitter2>()
+				},
+				{
+					provide: ConfigService,
+					useValue: createMock<ConfigService>()
+				}
+			]
+		}).compile();
+
+		useCase = module.get(UpdateBoardPhaseUseCase);
+		boardRepositoryMock = module.get(Boards.TYPES.repositories.BoardRepository);
+		eventEmitterMock = module.get(EventEmitter2);
+		configServiceMock = module.get(ConfigService);
+		slackSendMessageServiceMock = module.get(
+			CommunicationsType.TYPES.services.SlackSendMessageService
+		);
+	});
+
+	beforeEach(() => {
+		jest.restoreAllMocks();
+		jest.clearAllMocks();
+	});
+
+	it('should be defined', () => {
+		expect(useCase).toBeDefined();
+	});
+
+	describe('updatePhase', () => {
+		it('should call the boardRepository ', async () => {
+			const boardPhaseDto = { boardId: '6405f9a04633b1668f71c068', phase: BoardPhases.ADDCARDS };
+			await useCase.execute(boardPhaseDto);
+			expect(boardRepositoryMock.updatePhase).toBeCalledTimes(1);
+		});
+
+		it('should throw the badRequestException when the boardRepository fails', async () => {
+			const boardPhaseDto = { boardId: '6405f9a04633b1668f71c068', phase: BoardPhases.ADDCARDS };
+			// Set up the board repository mock to reject with an error
+			boardRepositoryMock.updatePhase.mockRejectedValueOnce(new Error('Some error'));
+
+			// Verify that the service method that is being tested throws a BadRequestException
+			expect(async () => await useCase.execute(boardPhaseDto)).rejects.toThrowError(
+				UpdateFailedException
+			);
+		});
+
+		it('should call the websocket with eventEmitter', async () => {
+			const boardPhaseDto = { boardId: '6405f9a04633b1668f71c068', phase: BoardPhases.ADDCARDS };
+			// Call the service method being tested
+			await useCase.execute(boardPhaseDto);
+
+			// Verify that the eventEmitterMock.emit method was called exactly once
+			expect(eventEmitterMock.emit).toHaveBeenCalledTimes(1);
+		});
+
+		it('should call the slackSendMessageService.execute once with slackMessageDto', async () => {
+			const boardPhaseDto = { boardId: '6405f9a04633b1668f71c068', phase: BoardPhases.ADDCARDS };
+			// Create a fake board object with the specified properties
+			const board = BoardFactory.create();
+			board.team = TeamFactory.create({ name: 'xgeeks' });
+
+			board.phase = BoardPhases.SUBMITTED;
+			board.slackEnable = true;
+
+			const table = {
+				[SLACK_MASTER_CHANNEL_ID]: '6405f9a04633b1668f71c068',
+				[SLACK_ENABLE]: true,
+				[FRONTEND_URL]: 'https://split.kigroup.de/'
+			};
+
+			// Set up the board repository mock to resolve with the fake board object
+			boardRepositoryMock.updatePhase.mockResolvedValue(board);
+
+			// Set up the configuration service mock
+			configServiceMock.getOrThrow.mockImplementation((key: string) => {
+				return table[key];
+			});
+
+			// Call the service method being tested
+			await useCase.execute(boardPhaseDto);
+
+			// Verify that the slackSendMessageService.execute method with correct data 1 time
+			expect(slackSendMessageServiceMock.execute).toHaveBeenNthCalledWith(1, {
+				slackChannelId: '6405f9a04633b1668f71c068',
+				message: expect.stringContaining('https://split.kigroup.de/')
+			});
+
+			board.phase = BoardPhases.VOTINGPHASE;
+
+			await useCase.execute(boardPhaseDto);
+
+			// Verify that the slackSendMessageService.execute method with correct data 1 time
+			expect(slackSendMessageServiceMock.execute).toHaveBeenNthCalledWith(1, {
+				slackChannelId: '6405f9a04633b1668f71c068',
+				message: expect.stringContaining('https://split.kigroup.de/')
+			});
+		});
+	});
+});
